Fix TranscriptFetcher BIGINT-as-string deserialization gap
node-postgres deserializes BIGINT/int8 columns as JS strings by default,
so start_ts was silently coming back as a string despite TranscriptEventRow
declaring it as number. Cast to ::int in the query (meeting-relative ms
timestamps comfortably fit in 32 bits) and add a regression assertion.

+++ src/knowledgeGraph/transcriptFetcher.ts
@@ -5,7 +5,7 @@ import type { FormattedTranscript } from "./knowledgeGraph.types";
 export class TranscriptFetcher {
   async fetchFormattedTranscript(meetingId: string): Promise<FormattedTranscript> {
     const { rows } = await getPool().query(
-      `SELECT participant_id AS "participantId", speaker_name AS "speakerName", text, start_ts AS "startTs"
+      `SELECT participant_id AS "participantId", speaker_name AS "speakerName", text, start_ts::int AS "startTs"
        FROM transcript_events
        WHERE meeting_id = $1
        ORDER BY sequence_number`,
